making changes

+++ application/migrations/610_version_610.php
@@ -10,10 +10,15 @@ class Migration_Version_610 extends CI_Migration
 
     public function up()
     {
-        $this->db->query("INSERT INTO `tbl_online_payment` (`online_payment_id`, `gateway_name`, `icon`, `field_1`, `field_2`, `field_3`, `field_4`, `field_5`, `link`, `modal`) VALUES 
-        (NULL, 'PipraPay', 'piprapay.png', 'piprapay_api_url', 'piprapay_api_key', 'piprapay_api_secret', 'piprapay_merchant_id', '', 'payment/piprapay', 'Yes');");
+        $piprapay_payment = $this->db->get_where('tbl_online_payment', array('gateway_name' => 'PipraPay'))->row();
+        if (empty($piprapay_payment)) {
+            $this->db->query("INSERT INTO `tbl_online_payment` (`online_payment_id`, `gateway_name`, `icon`, `field_1`, `field_2`, `field_3`, `field_4`, `field_5`, `link`, `modal`) VALUES 
+            (NULL, 'PipraPay', 'piprapay.png', 'piprapay_api_url', 'piprapay_api_key', 'piprapay_api_secret', 'piprapay_merchant_id', '', 'payment/piprapay', 'Yes');");
+        }
 
-        $this->db->query("ALTER TABLE `tbl_invoices` ADD `allow_piprapay` ENUM('Yes','No') NULL DEFAULT 'Yes' AFTER `allow_tappayment`;");
+        if (!$this->db->field_exists('allow_piprapay', 'tbl_invoices')) {
+            $this->db->query("ALTER TABLE `tbl_invoices` ADD `allow_piprapay` ENUM('Yes','No') NULL DEFAULT 'Yes' AFTER `allow_tappayment`;");
+        }
 
         $config_items = array(
             array(
@@ -83,8 +88,11 @@ class Migration_Version_610 extends CI_Migration
         );
 
         foreach ($config_items as $item) {
-            $this->db->query("INSERT INTO `tbl_config` (`config_key`, `value`, `label`, `type`, `options`, `description`) VALUES 
-            ('" . $item['config_key'] . "', '" . $item['value'] . "', '" . $item['label'] . "', '" . $item['type'] . "', '" . $item['options'] . "', '" . $item['description'] . "');");
+            $check_config = $this->db->get_where('tbl_config', array('config_key' => $item['config_key']))->row();
+            if (empty($check_config)) {
+                $this->db->query("INSERT INTO `tbl_config` (`config_key`, `value`, `label`, `type`, `options`, `description`) VALUES 
+                ('" . $item['config_key'] . "', '" . $item['value'] . "', '" . $item['label'] . "', '" . $item['type'] . "', '" . $item['options'] . "', '" . $item['description'] . "');");
+            }
         }
 
         $this->db->query("UPDATE `tbl_config` SET `value` = '6.1.0' WHERE `tbl_config`.`config_key` = 'version';");
@@ -109,7 +117,9 @@ class Migration_Version_610 extends CI_Migration
             $this->db->query("DELETE FROM `tbl_config` WHERE `config_key` = '" . $key . "';");
         }
 
-        $this->db->query("ALTER TABLE `tbl_invoices` DROP COLUMN `allow_piprapay`;");
+        if ($this->db->field_exists('allow_piprapay', 'tbl_invoices')) {
+            $this->db->query("ALTER TABLE `tbl_invoices` DROP COLUMN `allow_piprapay`;");
+        }
 
         $this->db->query("UPDATE `tbl_config` SET `value` = '6.0.7' WHERE `tbl_config`.`config_key` = 'version';");
     }
